chore: util 파일 관심사별로 분리
