+++ src/utils/couponTimeString.ts
@@ -1,20 +1,3 @@
-import { ITEM_CHECKING_STATE_KEY } from "../constants";
-import { getStorage, setStorage } from "./localStorage/localStorage";
-
-export const API_TOKEN = `Basic ${btoa(`${process.env.VITE_API_USER_ID}:${process.env.VITE_API_USER_PASSWORD}`)}`;
-
-export const setCartItemCheckedStateInStorage = (id: number, isCheck: boolean) => {
-  const localData = getStorage<CartItemCheckedStateInStorage>(ITEM_CHECKING_STATE_KEY, {});
-  const newData = { ...localData, [id]: isCheck };
-  setStorage(ITEM_CHECKING_STATE_KEY, newData);
-};
-
-export const deleteCartItemCheckedStateInStorage = (id: number) => {
-  const localData = getStorage<CartItemCheckedStateInStorage>(ITEM_CHECKING_STATE_KEY, {});
-  delete localData[id];
-  setStorage(ITEM_CHECKING_STATE_KEY, localData);
-};
-
 // hh:mm:ss -> 오전/오후 h시 m분 s초
 export const changeTimeStringToKorean = (timeString: string, withPeriod: boolean) => {
   const [hours, minutes, seconds] = timeString.split(":").map(Number);
